Split Android live smoke support packages
