+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/AppStateResetRule.kt
@@ -5,8 +5,8 @@ import androidx.core.app.NotificationManagerCompat
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
-import com.flashcardsopensourceapp.app.livesmoke.currentBlockingSystemDialogSummaryOrNull
-import com.flashcardsopensourceapp.app.livesmoke.dismissBlockingSystemDialogIfPresent
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.currentBlockingSystemDialogSummaryOrNull
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.dismissBlockingSystemDialogIfPresent
 import com.flashcardsopensourceapp.app.prompts.guestreview.guestSignInAfterReviewPromptPreferencesName
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeout
