+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/NotificationTapSmokeTest.kt
@@ -9,16 +9,16 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForFlowValue
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilAtLeastOneExistsOrFail
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilWithMitigation
+import com.flashcardsopensourceapp.app.livesmoke.flows.openCardsTab
+import com.flashcardsopensourceapp.app.livesmoke.support.LiveSmokeContext
+import com.flashcardsopensourceapp.app.livesmoke.support.appGraph
+import com.flashcardsopensourceapp.app.livesmoke.support.externalUiTimeoutMillis
+import com.flashcardsopensourceapp.app.livesmoke.support.internalUiTimeoutMillis
+import com.flashcardsopensourceapp.app.livesmoke.support.step
 import com.flashcardsopensourceapp.core.ui.VisibleAppScreen
-import com.flashcardsopensourceapp.app.livesmoke.LiveSmokeContext
-import com.flashcardsopensourceapp.app.livesmoke.appGraph
-import com.flashcardsopensourceapp.app.livesmoke.externalUiTimeoutMillis
-import com.flashcardsopensourceapp.app.livesmoke.internalUiTimeoutMillis
-import com.flashcardsopensourceapp.app.livesmoke.openCardsTab
-import com.flashcardsopensourceapp.app.livesmoke.step
-import com.flashcardsopensourceapp.app.livesmoke.waitForFlowValue
-import com.flashcardsopensourceapp.app.livesmoke.waitUntilAtLeastOneExistsOrFail
-import com.flashcardsopensourceapp.app.livesmoke.waitUntilWithMitigation
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
