+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/NotificationTapSmokeTestSupport.kt
@@ -9,17 +9,17 @@ import androidx.core.app.NotificationManagerCompat
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.By
 import androidx.test.uiautomator.Until
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.countNodesWithTagInAnySemanticsTree
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.currentBlockingSystemDialogSummaryOrNull
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.dismissExternalSystemDialogIfPresent
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForFlowValue
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilWithMitigation
+import com.flashcardsopensourceapp.app.livesmoke.support.LiveSmokeContext
+import com.flashcardsopensourceapp.app.livesmoke.support.appGraph
+import com.flashcardsopensourceapp.app.livesmoke.support.externalUiTimeoutMillis
+import com.flashcardsopensourceapp.app.livesmoke.support.internalUiTimeoutMillis
 import com.flashcardsopensourceapp.app.notifications.hasNotificationPermission
 import com.flashcardsopensourceapp.app.notifications.showReviewReminderNotification
-import com.flashcardsopensourceapp.app.livesmoke.LiveSmokeContext
-import com.flashcardsopensourceapp.app.livesmoke.appGraph
-import com.flashcardsopensourceapp.app.livesmoke.countNodesWithTagInAnySemanticsTree
-import com.flashcardsopensourceapp.app.livesmoke.currentBlockingSystemDialogSummaryOrNull
-import com.flashcardsopensourceapp.app.livesmoke.dismissExternalSystemDialogIfPresent
-import com.flashcardsopensourceapp.app.livesmoke.externalUiTimeoutMillis
-import com.flashcardsopensourceapp.app.livesmoke.internalUiTimeoutMillis
-import com.flashcardsopensourceapp.app.livesmoke.waitForFlowValue
-import com.flashcardsopensourceapp.app.livesmoke.waitUntilWithMitigation
 import com.flashcardsopensourceapp.core.ui.VisibleAppScreen
 import com.flashcardsopensourceapp.feature.review.reviewCurrentCardTag
 import com.flashcardsopensourceapp.feature.review.reviewEmptyStateTag
