+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/flows/LiveSmokeWorkspaceProgressFlows.kt
@@ -1,12 +1,22 @@
 @file:OptIn(androidx.compose.ui.test.ExperimentalTestApi::class)
 
-package com.flashcardsopensourceapp.app.livesmoke
+package com.flashcardsopensourceapp.app.livesmoke.flows
 
 import androidx.compose.ui.test.hasClickAction
 import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performTextReplacement
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickNode
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickTag
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.nodeSummary
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForTagToDisappear
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForTagToExist
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilWithMitigation
+import com.flashcardsopensourceapp.app.livesmoke.support.LiveSmokeContext
+import com.flashcardsopensourceapp.app.livesmoke.support.currentCloudSettingsSummary
+import com.flashcardsopensourceapp.app.livesmoke.support.currentWorkspaceSummaryOrNull
+import com.flashcardsopensourceapp.app.livesmoke.support.externalUiTimeoutMillis
 import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewTodayDueCountTag
 import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewTodayNewCountTag
 import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewTodayReviewedCountTag
