+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/support/LiveSmokeAppGraphSupport.kt
@@ -1,0 +1,8 @@
+package com.flashcardsopensourceapp.app.livesmoke.support
+
+import com.flashcardsopensourceapp.app.FlashcardsApplication
+import com.flashcardsopensourceapp.app.di.AppGraph
+
+internal fun LiveSmokeContext.appGraph(): AppGraph {
+    return (composeRule.activity.application as FlashcardsApplication).appGraph
+}
