+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/support/LiveSmokeScenarioAccountAssertions.kt
@@ -1,0 +1,33 @@
+package com.flashcardsopensourceapp.app.livesmoke.support
+
+import androidx.compose.ui.test.hasClickAction
+import androidx.compose.ui.test.hasText
+import androidx.compose.ui.test.onNodeWithText
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickNode
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickText
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.tapBackIcon
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilAtLeastOneExistsOrFail
+import com.flashcardsopensourceapp.app.livesmoke.flows.openSettingsTab
+
+internal fun LiveSmokeContext.assertLinkedAccountStatus(
+    reviewEmail: String,
+    workspaceName: String
+) {
+    openSettingsTab()
+    composeRule.onNodeWithText("Current Workspace").fetchSemanticsNode()
+    clickNode(
+        matcher = hasText("Account").and(other = hasClickAction()),
+        label = "Account"
+    )
+    clickText(text = "Account status", substring = false)
+    waitUntilAtLeastOneExistsOrFail(
+        matcher = hasText(reviewEmail),
+        timeoutMillis = internalUiTimeoutMillis
+    )
+    composeRule.onNodeWithText("Linked").fetchSemanticsNode()
+    tapBackIcon()
+    composeRule.onNodeWithText("Account Settings").fetchSemanticsNode()
+    tapBackIcon()
+    composeRule.onNodeWithText("Current Workspace").fetchSemanticsNode()
+    composeRule.onNodeWithText(workspaceName).fetchSemanticsNode()
+}
