+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/support/LiveSmokeScenarioAiHelpers.kt
@@ -1,34 +1,36 @@
-package com.flashcardsopensourceapp.app.livesmoke
+package com.flashcardsopensourceapp.app.livesmoke.support
 
 import androidx.compose.ui.semantics.SemanticsProperties
 import androidx.compose.ui.semantics.getOrNull
 import androidx.compose.ui.test.hasContentDescription
-import androidx.compose.ui.test.hasClickAction
 import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.onAllNodesWithTag
-import androidx.compose.ui.test.onAllNodesWithText
-import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextClearance
 import androidx.compose.ui.test.performTextInput
 import androidx.compose.ui.test.performTextReplacement
-import com.flashcardsopensourceapp.app.RepositorySeedCard
-import com.flashcardsopensourceapp.app.RepositorySeedScenario
-import com.flashcardsopensourceapp.app.createRepositorySeedExecutor
+import com.flashcardsopensourceapp.app.di.AppGraph
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickTag
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.countNodesWithTagInAnySemanticsTree
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.currentBlockingSystemDialogSummaryOrNull
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.dismissExternalSystemDialogIfPresent
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.hasVisibleText
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.nodeSummary
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForFlowValue
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForTagToDisappear
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitForTagToExist
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilAtLeastOneExistsOrFail
+import com.flashcardsopensourceapp.app.livesmoke.diagnostics.waitUntilWithMitigation
+import com.flashcardsopensourceapp.app.livesmoke.flows.dismissAiConsentIfNeeded
+import com.flashcardsopensourceapp.app.livesmoke.flows.openAiTab
 import com.flashcardsopensourceapp.data.local.ai.store.AiChatHistoryStore
 import com.flashcardsopensourceapp.data.local.ai.store.makeAiChatHistoryScopedWorkspaceId
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatContentPart
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatPersistedState
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatRole
 import com.flashcardsopensourceapp.data.local.model.ai.AiChatToolCallStatus
-import com.flashcardsopensourceapp.data.local.model.cards.CardFilter
-import com.flashcardsopensourceapp.data.local.model.cards.CardSummary
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudAccountState
-import com.flashcardsopensourceapp.data.local.model.scheduling.EffortLevel
-import com.flashcardsopensourceapp.data.local.model.sync.SyncStatus
 import com.flashcardsopensourceapp.feature.ai.aiAssistantMessageBubbleTag
 import com.flashcardsopensourceapp.feature.ai.aiAssistantTextPartTag
 import com.flashcardsopensourceapp.feature.ai.aiComposerMessageFieldTag
@@ -39,181 +41,8 @@ import com.flashcardsopensourceapp.feature.ai.aiEmptyStateTag
 import com.flashcardsopensourceapp.feature.ai.aiNewChatButtonTag
 import com.flashcardsopensourceapp.feature.ai.aiUserMessageBubbleTag
 import com.flashcardsopensourceapp.feature.ai.R as AiFeatureR
-import com.flashcardsopensourceapp.feature.cards.cardsCardFrontTextTag
-import com.flashcardsopensourceapp.feature.review.reviewCurrentCardFrontContentTag
-import com.flashcardsopensourceapp.feature.review.reviewEmptyStateTitleTag
-import com.flashcardsopensourceapp.feature.review.reviewRateGoodButtonTag
-import com.flashcardsopensourceapp.feature.review.reviewShowAnswerButtonTag
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
-
-internal fun LiveSmokeContext.withLinkedWorkspaceSession(
-    reviewEmail: String,
-    workspaceName: String,
-    action: () -> Unit
-) {
-    var primaryFailure: Throwable? = null
-    var ephemeralWorkspaceHandle: EphemeralWorkspaceHandle? = null
-
-    try {
-        step("sign in with the configured review account") {
-            signInWithReviewAccount(reviewEmail = reviewEmail)
-        }
-        step("create an isolated linked workspace for this run") {
-            ephemeralWorkspaceHandle = createEphemeralWorkspace(workspaceName = workspaceName)
-        }
-        action()
-    } catch (error: Throwable) {
-        primaryFailure = error
-        throw error
-    } finally {
-        if (ephemeralWorkspaceHandle != null) {
-            if (primaryFailure != null) {
-                resetInlineRawScreenStateFailureGuard()
-            }
-            try {
-                step("delete the isolated workspace") {
-                    deleteEphemeralWorkspace(workspaceHandle = requireNotNull(ephemeralWorkspaceHandle))
-                }
-            } catch (cleanupError: Throwable) {
-                if (primaryFailure != null) {
-                    primaryFailure.addSuppressed(cleanupError)
-                } else {
-                    throw cleanupError
-                }
-            }
-        }
-    }
-}
-
-internal fun LiveSmokeContext.deleteCurrentWorkspaceCards() {
-    val deletedCards: List<CardSummary> = runBlocking {
-        val appGraph = appGraph()
-        val currentCards: List<CardSummary> = appGraph.cardsRepository.observeCards(
-            searchQuery = "",
-            filter = CardFilter(
-                tags = emptyList(),
-                effort = emptyList()
-            )
-        ).first()
-        if (currentCards.isEmpty()) {
-            return@runBlocking emptyList<CardSummary>()
-        }
-
-        currentCards.forEach { card ->
-            appGraph.cardsRepository.deleteCard(cardId = card.cardId)
-        }
-        appGraph.syncRepository.syncNow()
-
-        val syncStatus = appGraph.syncRepository.observeSyncStatus().first()
-        if (syncStatus.status != SyncStatus.Idle || syncStatus.lastErrorMessage.isNotEmpty()) {
-            throw AssertionError(
-                "Workspace card cleanup sync did not finish cleanly. " +
-                    "Status=${syncStatus.status} " +
-                    "LastError=${syncStatus.lastErrorMessage} " +
-                    "CloudSettings=${currentCloudSettingsSummary()} " +
-                    "CurrentWorkspace=${currentWorkspaceSummaryOrNull()}"
-            )
-        }
-
-        val remainingCards: List<CardSummary> = appGraph.cardsRepository.observeCards(
-            searchQuery = "",
-            filter = CardFilter(
-                tags = emptyList(),
-                effort = emptyList()
-            )
-        ).first()
-        if (remainingCards.isNotEmpty()) {
-            throw AssertionError(
-                "Workspace card cleanup left active cards behind. " +
-                    "RemainingCards=${remainingCards.map { card -> card.cardId }} " +
-                    "CloudSettings=${currentCloudSettingsSummary()} " +
-                    "CurrentWorkspace=${currentWorkspaceSummaryOrNull()}"
-            )
-        }
-
-        currentCards
-    }
-
-    if (deletedCards.isNotEmpty()) {
-        composeRule.waitForIdle()
-    }
-}
-
-internal fun LiveSmokeContext.createManualCard(
-    frontText: String,
-    backText: String,
-    markerTag: String
-) {
-    openCardsTab()
-    clickContentDescription(contentDescription = "Add card")
-    updateCardText(fieldTitle = "Front", value = frontText)
-    updateCardText(fieldTitle = "Back", value = backText)
-    clickText(text = "Tags", substring = false)
-    composeRule.onNodeWithText("Add a tag").performTextInput(markerTag)
-    clickText(text = "Add tag", substring = false)
-    tapBackIcon()
-    scrollToText(text = "Save")
-    waitUntilAtLeastOneExistsOrFail(
-        matcher = hasClickAction().and(other = hasText("Save")),
-        timeoutMillis = internalUiTimeoutMillis
-    )
-    clickNode(
-        matcher = hasClickAction().and(other = hasText("Save")),
-        label = "Save card"
-    )
-    waitForTextToExist(
-        text = "Search cards",
-        substring = false,
-        timeoutMillis = internalUiTimeoutMillis,
-        context = "while waiting for cards search after saving a manual card"
-    )
-    waitUntilWithMitigation(
-        timeoutMillis = internalUiTimeoutMillis,
-        context = "while waiting for the saved manual card to appear"
-    ) {
-        visibleCardsFrontTexts().any { text -> text == frontText }
-    }
-}
-
-internal fun LiveSmokeContext.rateVisibleReviewCardGood() {
-    waitUntilAtLeastOneExistsOrFail(
-        matcher = hasText("Show answer"),
-        timeoutMillis = internalUiTimeoutMillis
-    )
-    clickTag(tag = reviewShowAnswerButtonTag, label = "Show answer")
-    clickTag(tag = reviewRateGoodButtonTag, label = "Rate Good")
-    waitUntilWithMitigation(
-        timeoutMillis = internalUiTimeoutMillis,
-        context = "while waiting for the review queue to advance"
-    ) {
-        composeRule.onAllNodesWithTag(reviewShowAnswerButtonTag).fetchSemanticsNodes().isNotEmpty() ||
-            composeRule.onAllNodesWithText("Session complete").fetchSemanticsNodes().isNotEmpty()
-    }
-}
-
-internal fun LiveSmokeContext.seedCardViaRepository(
-    frontText: String,
-    backText: String,
-    markerTag: String
-) {
-    runBlocking {
-        createRepositorySeedExecutor().seedCardsAndReviewsInCurrentWorkspace(
-            seedScenario = RepositorySeedScenario(
-                cards = listOf(
-                    RepositorySeedCard(
-                        frontText = frontText,
-                        backText = backText,
-                        tags = listOf(markerTag),
-                        effortLevel = EffortLevel.MEDIUM,
-                        reviews = emptyList()
-                    )
-                )
-            )
-        )
-    }
-    composeRule.waitForIdle()
-}
 
 private const val aiCreatePromptText: String =
     "I give you all permissions. Please create one test flashcard now."
@@ -390,74 +219,6 @@ internal fun LiveSmokeContext.createGuestAiConversationForReset() {
     )
 }
 
-internal fun LiveSmokeContext.assertCardVisibleInCards(searchText: String, timeoutMillis: Long) {
-    openCardsTab()
-    composeRule.onNodeWithText("Search cards").performTextReplacement(searchText)
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = timeoutMillis,
-            context = "while waiting for cards to show '$searchText'"
-        ) {
-            visibleCardsFrontTexts().any { text -> text == searchText }
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Cards did not show '$searchText'. " +
-                "VisibleCardFronts=${visibleCardsFrontTexts()} " +
-                "LocalCard=${localCardSnapshotOrNull(expectedFrontText = searchText)} " +
-                "SystemDialog=${currentBlockingSystemDialogSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-internal fun LiveSmokeContext.assertCardReachableInReview(
-    expectedFrontText: String,
-    timeoutMillis: Long
-) {
-    openReviewTab()
-    try {
-        waitUntilWithMitigation(
-            timeoutMillis = timeoutMillis,
-            context = "while waiting for review to show '$expectedFrontText'"
-        ) {
-            currentReviewCardFrontTextOrNull()?.contains(other = expectedFrontText) == true
-        }
-    } catch (error: Throwable) {
-        throw AssertionError(
-            "Review did not show '$expectedFrontText'. " +
-                "CurrentReviewFront=${currentReviewCardFrontTextOrNull()} " +
-                "ReviewEmptyStateTitle=${reviewEmptyStateTitleOrNull()} " +
-                "LocalCard=${localCardSnapshotOrNull(expectedFrontText = expectedFrontText)} " +
-                "SystemDialog=${currentBlockingSystemDialogSummaryOrNull()}",
-            error
-        )
-    }
-}
-
-internal fun LiveSmokeContext.assertLinkedAccountStatus(
-    reviewEmail: String,
-    workspaceName: String
-) {
-    openSettingsTab()
-    composeRule.onNodeWithText("Current Workspace").fetchSemanticsNode()
-    clickNode(
-        matcher = hasText("Account").and(other = hasClickAction()),
-        label = "Account"
-    )
-    clickText(text = "Account status", substring = false)
-    waitUntilAtLeastOneExistsOrFail(
-        matcher = hasText(reviewEmail),
-        timeoutMillis = internalUiTimeoutMillis
-    )
-    composeRule.onNodeWithText("Linked").fetchSemanticsNode()
-    tapBackIcon()
-    composeRule.onNodeWithText("Account Settings").fetchSemanticsNode()
-    tapBackIcon()
-    composeRule.onNodeWithText("Current Workspace").fetchSemanticsNode()
-    composeRule.onNodeWithText(workspaceName).fetchSemanticsNode()
-}
-
 private fun LiveSmokeContext.waitForAiComposerEditable(context: String) {
     try {
         waitUntilWithMitigation(
@@ -614,7 +375,7 @@ private fun LiveSmokeContext.waitForGuestAiEntryReady(
 private fun LiveSmokeContext.currentGuestAiEntryReadinessSnapshot(
     expectedLabel: String
 ): GuestAiEntryReadinessSnapshot {
-    val appGraph = appGraph()
+    val appGraph: AppGraph = appGraph()
     val cloudSettings = runBlocking {
         appGraph.cloudAccountRepository.observeCloudSettings().first()
     }
@@ -917,46 +678,6 @@ private fun LiveSmokeContext.latestAssistantMessageTextOrNull(): String? {
 
 private fun LiveSmokeContext.aiSendLabel(): String {
     return composeRule.activity.getString(AiFeatureR.string.ai_send)
-}
-
-private fun LiveSmokeContext.visibleCardsFrontTexts(): List<String> {
-    return composeRule.onAllNodesWithTag(cardsCardFrontTextTag, useUnmergedTree = true)
-        .fetchSemanticsNodes()
-        .map(::nodeSummary)
-        .filter { text -> text.isNotBlank() }
-}
-
-private fun LiveSmokeContext.currentReviewCardFrontTextOrNull(): String? {
-    return composeRule.onAllNodesWithTag(reviewCurrentCardFrontContentTag, useUnmergedTree = true)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-}
-
-private fun LiveSmokeContext.reviewEmptyStateTitleOrNull(): String? {
-    return composeRule.onAllNodesWithTag(reviewEmptyStateTitleTag, useUnmergedTree = true)
-        .fetchSemanticsNodes()
-        .singleOrNull()
-        ?.let(::nodeSummary)
-}
-
-private fun LiveSmokeContext.localCardSnapshotOrNull(expectedFrontText: String): String? {
-    return runBlocking {
-        val database = appGraph().database
-        val matchingCard = database.cardDao()
-            .observeCardsWithRelations()
-            .first()
-            .lastOrNull { cardWithRelations ->
-                cardWithRelations.card.frontText == expectedFrontText
-            } ?: return@runBlocking null
-        "cardId=${matchingCard.card.cardId} " +
-            "workspaceId=${matchingCard.card.workspaceId} " +
-            "dueAtMillis=${matchingCard.card.dueAtMillis} " +
-            "fsrsCardState=${matchingCard.card.fsrsCardState} " +
-            "reps=${matchingCard.card.reps} " +
-            "lapses=${matchingCard.card.lapses} " +
-            "tags=${matchingCard.tags.map { tag -> tag.name }}"
-    }
 }
 
 private fun LiveSmokeContext.currentWorkspaceIdOrThrow(context: String): String {
